moving groupBy util function to utils folder
