+++ app/utils/groupBy.ts
@@ -1,0 +1,10 @@
+export const groupBy =
+  (key: string) =>
+  <T>(array: T[]) =>
+    array.reduce(
+      (objectsByKeyValue, obj) => ({
+        ...objectsByKeyValue,
+        [obj[key]]: (objectsByKeyValue[obj[key]] || []).concat(obj),
+      }),
+      {},
+    )
